+++ utility/base_funs/funs.go
@@ -3,13 +3,67 @@ package base_funs
 import (
 	"context"
 	"fmt"
-	"github.com/gogf/gf/v2/container/garray"
-	"github.com/gogf/gf/v2/text/gstr"
-	"github.com/kysion/base-library/base_model"
+	"math/rand"
 	"reflect"
 	"sync"
 	"time"
+
+	"github.com/gogf/gf/v2/container/garray"
+	"github.com/gogf/gf/v2/text/gstr"
+	"github.com/kysion/base-library/base_model"
 )
+
+// RandomInt 生成一个介于 min 和 max 之间的随机整数
+// min: 最小值
+// max: 最大值
+// 返回值: 一个介于 min 和 max 之间的随机整数
+func RandomInt(min, max int) int {
+	source := rand.NewSource(time.Now().UnixNano())
+	r := rand.New(source)
+	return min + r.Intn(max-min+1)
+}
+
+// SetInterval 模拟 JavaScript 中的 setInterval 函数
+// interval 是执行任务的时间间隔
+// fn 是要周期性执行的函数
+// 返回一个通道，通过关闭该通道可以停止定时器
+func SetInterval(interval time.Duration, fn func()) chan struct{} {
+	stop := make(chan struct{})
+	ticker := time.NewTicker(interval)
+
+	go func() {
+		for {
+			select {
+			case <-ticker.C:
+				// 检查当前时间是否大于 21 点
+				currentHour := time.Now().Hour()
+				currentMinute := time.Now().Minute()
+				if currentHour >= 21 && currentMinute >= 25 {
+					ticker.Stop()
+					close(stop)
+					return
+				}
+				fn()
+			case <-stop:
+				ticker.Stop()
+				return
+			}
+		}
+	}()
+
+	return stop
+}
+
+// Ternary 实现类似三元操作符的功能
+// condition: 条件表达式
+// trueVal: 条件为真时返回的值
+// falseVal: 条件为假时返回的值
+func If[T any](condition bool, trueVal, falseVal T) T {
+	if condition {
+		return trueVal
+	}
+	return falseVal
+}
 
 // Contains 检查给定的切片中是否包含指定的元素。
 //
@@ -27,7 +81,7 @@ import (
 //	该函数使用泛型定义，支持多种类型的操作，提高了代码的通用性和复用性。
 func Contains[T comparable](slice []T, element T) bool {
 	// 检查切片是否为空或长度为0，如果是，则直接返回 false，因为切片中不可能包含指定元素。
-	if slice == nil || len(slice) == 0 {
+	if len(slice) == 0 {
 		return false
 	}
 
@@ -84,14 +138,6 @@ func FilterEmpty(slice []string) []string {
 	}
 	// 返回最终的结果切片。
 	return result
-}
-
-func If[R any](condition bool, trueVal, falseVal R) R {
-	if condition {
-		return trueVal
-	} else {
-		return falseVal
-	}
 }
 
 // SearchFilterEx 支持增加拓展字段，会提炼拓展字段的最新过滤题条件模型返回，并支持从原始的过滤模型剔除不需要的条件