+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-## [](https://github.com/kysion/base-library/compare/v0.3.3...v) (2025-04-15)
+## [](https://github.com/kysion/base-library/compare/v0.3.3...v) (2025-04-17)
 
 ### Features
 
+* **base_funs:** 添加 Ternary 函数实现三元操作符功能，并优化 Contains 函数逻辑 ([e360da1](https://github.com/kysion/base-library/commit/e360da16e2fefaa6bbcee9832e87718ac1858511))
 * **base_model:** 在 FilterInfo 结构中添加子查询字段，并优化查询逻辑 ([064c052](https://github.com/kysion/base-library/commit/064c05274c7fd803bc98c28773d8e2061178304c))
 ## [0.3.2](https://github.com/kysion/base-library/compare/v0.3.1...v0.3.2) (2025-04-08)